✨ 新機能(interpreter.rs): 単項演算子のサポートを追加 ♻️ リファクタリング(lexical.rs): 整数トークンの型をi64からi32に変更 ♻️ リファクタリング(main.rs): コメントアウトされたコードを有効化し、デバッグ出力を削除

+++ core/src/lexical.rs
@@ -124,7 +124,7 @@ impl Token
 
             // 数値の場合
             _ if keyword.parse::<i64>().is_ok() =>
-                Some(Token::Constant(Constant::Integer(keyword.parse::<i64>().unwrap()))),
+                Some(Token::Constant(Constant::Integer(keyword.parse::<i32>().unwrap()))),
             _ if keyword.parse::<f64>().is_ok() =>
                 Some(Token::Constant(Constant::Float(keyword.parse::<f64>().unwrap()))),
             _ => None,
